pyproject.toml
[tool.poetry]
name = "frontegg"
version = "2.0.0"
description = "Frontegg is a web platform where SaaS companies can set up their fully managed, scalable and brand aware - SaaS features and integrate them into their SaaS portals in up to 5 lines of code."
homepage = "https://frontegg.com/"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Operating System :: OS Independent"
]
authors = ["Frontegg LTD <hello@frontegg.com>"]
readme = "README-PYPI.rst"

[tool.poetry.dependencies]
python = "^3.7"
flask = { version = "^1.0", optional = true }
fastapi = { version = "*", optional = true }
requests = "^2.28.2"
arrow = "^1.2.3"
aiohttp = {version = "^3.6.2", optional = true}
PyJWT = "^2.6.0"
cryptography = "^39.0.0"

[tool.poetry.extras]
flask = ["flask"]
fastapi = ["fastapi"]

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"
pytest-flask = "^0.15.0"
pytest-vcr = "^1.0.2"
flake8-mypy = "^17.8.0"
flake8 = "^3.7.9"
autopep8 = "^1.4.4"
pydocstyle = "^5.0.2"

[tool.poetry.group.dev.dependencies]
flask-cors = "^3.0.10"
uvicorn = "^0.18.3"


[tool.poetry.group.extras.dependencies]
redis = "^4.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
